refactor profile config and keyring
