+++ connection_profile.go
@@ -2,15 +2,11 @@ package main
 
 import (
 	"errors"
-	"fmt"
 	"log"
 	"os"
 	"path/filepath"
 	"strconv"
 	"strings"
-
-	"github.com/BurntSushi/toml"
-	"github.com/zalando/go-keyring"
 
 	"github.com/marang/goemqutiti/internal/files"
 )
@@ -45,40 +41,6 @@ type Profile struct {
 	LastWillRetain      bool   `toml:"last_will_retain"`
 	LastWillPayload     string `toml:"last_will_payload"`
 	RandomIDSuffix      bool   `toml:"random_id_suffix"`
-}
-
-type Config struct {
-	DefaultProfile string    `toml:"default_profile"`
-	Profiles       []Profile `toml:"profiles"`
-}
-
-// DefaultUserConfigFile returns ~/.emqutiti/config.toml.
-func DefaultUserConfigFile() (string, error) {
-	home, err := os.UserHomeDir()
-	if err != nil {
-		return "", err
-	}
-	return filepath.Join(home, ".emqutiti", "config.toml"), nil
-}
-
-// RetrievePasswordFromKeyring resolves a keyring:<service>/<user> reference.
-func RetrievePasswordFromKeyring(password string) (string, error) {
-	if !strings.HasPrefix(password, "keyring:") {
-		return "", fmt.Errorf("password does not reference keyring")
-	}
-	parts := strings.SplitN(password, ":", 2)
-	if len(parts) != 2 {
-		return "", fmt.Errorf("invalid keyring reference: %s", password)
-	}
-	serviceUsername := strings.SplitN(parts[1], "/", 2)
-	if len(serviceUsername) != 2 {
-		return "", fmt.Errorf("invalid keyring format: %s", parts[1])
-	}
-	pw, err := keyring.Get(serviceUsername[0], serviceUsername[1])
-	if err != nil {
-		return "", fmt.Errorf("failed to retrieve password from keyring for %s/%s: %w", serviceUsername[0], serviceUsername[1], err)
-	}
-	return pw, nil
 }
 
 func sanitizeEnvName(name string) string {
@@ -222,93 +184,6 @@ func ApplyEnvVars(p *Profile) {
 			setter(p, val)
 		}
 	}
-}
-
-// LoadConfig reads profiles from a TOML file and resolves keyring references.
-func LoadConfig(filePath string) (*Config, error) {
-	var err error
-	if filePath == "" {
-		if filePath, err = DefaultUserConfigFile(); err != nil {
-			return nil, err
-		}
-	}
-	var cfg Config
-	if _, err := toml.DecodeFile(filePath, &cfg); err != nil {
-		return nil, fmt.Errorf("failed to decode config file: %w", err)
-	}
-	for i := range cfg.Profiles {
-		p := &cfg.Profiles[i]
-		if p.FromEnv {
-			ApplyEnvVars(p)
-			continue
-		}
-		if strings.HasPrefix(p.Password, "keyring:") {
-			pw, err := RetrievePasswordFromKeyring(p.Password)
-			if err != nil {
-				return nil, err
-			}
-			p.Password = pw
-		}
-	}
-	return &cfg, nil
-}
-
-// LoadProfile returns the named profile from the config file, falling back to the default or first profile.
-func LoadProfile(name, file string) (*Profile, error) {
-	cfg, err := LoadConfig(file)
-	if err != nil {
-		return nil, err
-	}
-	var p *Profile
-	if name != "" {
-		for i := range cfg.Profiles {
-			if cfg.Profiles[i].Name == name {
-				p = &cfg.Profiles[i]
-				break
-			}
-		}
-	} else if cfg.DefaultProfile != "" {
-		for i := range cfg.Profiles {
-			if cfg.Profiles[i].Name == cfg.DefaultProfile {
-				p = &cfg.Profiles[i]
-				break
-			}
-		}
-	}
-	if p == nil && len(cfg.Profiles) > 0 {
-		p = &cfg.Profiles[0]
-	}
-	if p == nil {
-		return nil, fmt.Errorf("no connection profile available")
-	}
-	return p, nil
-}
-
-// saveConfig persists profiles and default selection to config.toml.
-func saveConfig(profiles []Profile, defaultName string) {
-	saved := loadState()
-	cfg := userConfig{
-		DefaultProfileName: defaultName,
-		Profiles:           profiles,
-		Saved:              make(map[string]persistedConn),
-	}
-	for k, v := range saved {
-		var topics []persistedTopic
-		for _, t := range v.Topics {
-			topics = append(topics, persistedTopic{Title: t.title, Active: t.subscribed})
-		}
-		var payloads []persistedPayload
-		for _, p := range v.Payloads {
-			payloads = append(payloads, persistedPayload{Topic: p.topic, Payload: p.payload})
-		}
-		cfg.Saved[k] = persistedConn{Topics: topics, Payloads: payloads}
-	}
-	writeConfig(cfg)
-}
-
-// savePasswordToKeyring stores a password in the system keyring.
-func savePasswordToKeyring(service, username, password string) error {
-	return keyring.Set("emqutiti-"+service, username, password)
 }
 
 // deleteProfileData removes profile-specific persisted history and traces and
